+++ lib/ReactViews/Icon.jsx
@@ -116,7 +116,8 @@ export const GLYPHS = {
   uploadThin: require("../../wwwroot/images/icons/upload-thin.svg"),
   oneTwoThree: require("../../wwwroot/images/icons/one-two-three.svg"),
   accordionOpen: require("../../wwwroot/images/icons/accordion-open.svg"),
-  accordionClose: require("../../wwwroot/images/icons/accordion-close.svg")
+  accordionClose: require("../../wwwroot/images/icons/accordion-close.svg"),
+  editor: require("../../wwwroot/images/icons/editor.svg")
 };
 
 export const Icon = createReactClass({