+++ lib/ReactViews/Map/MapNavigation.jsx
@@ -19,9 +19,13 @@ import Icon from "../Icon";
 
 // import Icon from "../Icon";
 import Box from "../../Styled/Box";
+import Text from "../../Styled/Text";
 import MapIconButton from "../MapIconButton/MapIconButton";
 import FeedbackButton from "../Feedback/FeedbackButton";
 import CloseToolButton from "./Navigation/CloseToolButton";
+import Prompt from "../Generic/Prompt";
+import { runInAction } from "mobx";
+import { withTranslation } from "react-i18next";
 
 const StyledMapNavigation = styled.div`
   ${p =>
@@ -38,6 +42,7 @@ class MapNavigation extends React.Component {
     terria: PropTypes.object.isRequired,
     viewState: PropTypes.object.isRequired,
     theme: PropTypes.object.isRequired,
+    t: PropTypes.func.isRequired,
     navItems: PropTypes.arrayOf(PropTypes.element)
   };
 
@@ -46,7 +51,7 @@ class MapNavigation extends React.Component {
   };
 
   render() {
-    const { viewState } = this.props;
+    const { viewState, t } = this.props;
     const toolIsDifference =
       this.props.viewState.currentTool?.toolName === "Difference";
     const isDiffMode = this.props.viewState.isToolOpen && toolIsDifference;
@@ -135,10 +140,42 @@ class MapNavigation extends React.Component {
                     expandInPlace
                     iconElement={() => <Icon glyph={Icon.GLYPHS.helpThick} />}
                     onClick={() => this.props.viewState.showHelpPanel()}
+                    neverCollapse={
+                      this.props.viewState.featurePrompts.indexOf("help") >= 0
+                    }
                   >
                     Help
                   </MapIconButton>
                 </div>
+                <Prompt
+                  content={
+                    <div>
+                      <Text bold extraLarge textLight>
+                        {t("helpPanel.promptMessage")}
+                      </Text>
+                    </div>
+                  }
+                  displayDelay={500}
+                  dismissText={t("helpPanel.dismissText")}
+                  dismissAction={() => {
+                    runInAction(() =>
+                      this.props.viewState.toggleFeaturePrompt(
+                        "help",
+                        false,
+                        true
+                      )
+                    );
+                  }}
+                  caretTopOffset={75}
+                  caretLeftOffset={265}
+                  caretSize={15}
+                  promptWidth={273}
+                  promptTopOffset={-20}
+                  promptLeftOffset={-330}
+                  isVisible={
+                    this.props.viewState.featurePrompts.indexOf("help") >= 0
+                  }
+                />
               </If>
             </div>
           </Box>
@@ -148,4 +185,4 @@ class MapNavigation extends React.Component {
   }
 }
 
-export default withTheme(MapNavigation);
+export default withTranslation()(withTheme(MapNavigation));