+++ lib/ReactViews/Map/Panels/HelpPanel/TrainerPane.jsx
@@ -7,8 +7,6 @@ import styled from "styled-components";
 
 import StyledHtml from "./StyledHtml";
 
-import parseCustomMarkdownToReact from "../../../Custom/parseCustomMarkdownToReact";
-
 import Box, { BoxSpan } from "../../../../Styled/Box";
 import Button from "../../../../Styled/Button";
 import Spacing from "../../../../Styled/Spacing";
@@ -43,9 +41,7 @@ class TrainerPane extends React.Component {
       <Text textDark noFontSize>
         <Box column>
           {markdownText && (
-            <StyledHtml
-              content={parseCustomMarkdownToReact(markdownText).props.children}
-            />
+            <StyledHtml viewState={viewState} markdown={markdownText} />
           )}
           {trainerItems?.map && (
             <UlTrainerItems column fullWidth justifySpaceBetween>